Fixes a misuse of createRef
React.createRef cannot be used in functional components. Doing so here causes ref.current to sometimes be `null` when `measureSelf` is called (i.e. if a rerender has happened).

As stated in the React docs: "You may not use the ref attribute on function components because they don’t have instances." (https://reactjs.org/docs/refs-and-the-dom.html)

The solution proposed here is to use the `useRef` hook, which can be used in a functional component.