feat(calendar): add visibleDate prop to RangeCalendar
allows view to be adjusted when range is set outside
the currently displayed month/year.

previously when new date range was set outside of the
displayed month the view wouldn't update, and the user
would need to manually change the view to find the
selected range.

Uses UNSAFE_componentWillReceiveProps because it was
not possible to use getDerivedStateFromProps without
breaking the previous/next month buttons. This method
allows for the display to be set programmatically while
still allowing the user to page forward/back.